flask_app/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-distro
psycopg2
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests